ties user to article models (in POST endpoint) and API calls only return data owned by the user requesting the data

+++ routes/api/v1/jackets.js
@@ -4,8 +4,9 @@ const	express = require('express'),
 
 // Index
 router.get('/', async (req, res) => {
+	const { user } = req;
 	try {
-		const jackets = await Jacket.find({});
+		const jackets = await Jacket.find({owner: user._id});
 		res.json(jackets);
 	} catch (err) {
 		console.log(err);
@@ -15,6 +16,7 @@ router.get('/', async (req, res) => {
 
 // Detail
 router.get('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
 		const jacket = await Jacket.findById(id);
@@ -23,6 +25,9 @@ router.get('/:id', async (req, res) => {
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!jacket.owner.equals(user._id))
+			return res.sendStatus(403);
+
 		res.json(jacket);
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
@@ -36,9 +41,11 @@ router.get('/:id', async (req, res) => {
 
 // Create
 router.post('/', async (req, res) => {
+	const { user } = req;
 	try {
-		const { jacket } = req.body;
-		const newJacket = await Jacket.create(jacket);
+		const { jacket: jacketData } = req.body;
+		jacketData.owner = user._id;
+		const newJacket = await Jacket.create(jacketData);
 		res.json(newJacket);
 	} catch (err) {
 		console.log(err);
@@ -48,15 +55,21 @@ router.post('/', async (req, res) => {
 
 // Update
 router.put('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
-		const { jacket } = req.body;
-		const updatedJacket = await Jacket.findByIdAndUpdate(id, jacket, {new: true});
-		if (!updatedJacket) {
+		const { jacket: jacketData } = req.body;
+		const jacket = await Jacket.findById(id);
+
+		if (!jacket) {
 			const err = new Error('Jacket Not Found.');
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!jacket.owner.equals(user._id))
+			return res.sendStatus(403);
+
+		const updatedJacket = await Jacket.findByIdAndUpdate(id, jacketData, {new: true});
 		res.json(updatedJacket);
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
@@ -70,14 +83,20 @@ router.put('/:id', async (req, res) => {
 
 // Delete
 router.delete('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
-		const deletedJacket = await Jacket.findByIdAndRemove(id);
-		if (!deletedJacket) {
+		const jacket = await Jacket.findById(id);
+
+		if (!jacket) {
 			const err = new Error('Jacket Not Found.');
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!jacket.owner.equals(user._id))
+			return res.sendStatus(403);
+
+		await Jacket.findByIdAndRemove(id);
 		res.send('Successfully deleted.')
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
