+++ routes/api/v1/pants.js
@@ -4,8 +4,9 @@ const	express = require('express'),
 
 // Index
 router.get('/', async (req, res) => {
+	const { user } = req;
 	try {
-		const pants = await Pants.find({});
+		const pants = await Pants.find({owner: user._id});
 		res.json(pants);
 	} catch (err) {
 		console.log(err);
@@ -15,6 +16,7 @@ router.get('/', async (req, res) => {
 
 // Detail
 router.get('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
 		const pair = await Pants.findById(id);
@@ -23,6 +25,9 @@ router.get('/:id', async (req, res) => {
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!pair.owner.equals(user._id))
+			return res.sendStatus(403);
+
 		res.json(pair);
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
@@ -36,9 +41,11 @@ router.get('/:id', async (req, res) => {
 
 // Create
 router.post('/', async (req, res) => {
+	const { user } = req;
 	try {
-		const { pair } = req.body;
-		const newPair = await Pants.create(pair);
+		const { pair: pairData } = req.body;
+		pairData.owner = user._id
+		const newPair = await Pants.create(pairData);
 		res.json(newPair);
 	} catch (err) {
 		console.log(err);
@@ -48,15 +55,21 @@ router.post('/', async (req, res) => {
 
 // Update
 router.put('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
-		const { pair } = req.body;
-		const updatedPair = await Pants.findByIdAndUpdate(id, pair, {new: true});
-		if (!updatedPair) {
+		const { pair: pairData } = req.body;
+		const pair = await Pants.findById(id);
+
+		if (!pair) {
 			const err = new Error('Pants Not Found.');
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!pair.owner.equals(user._id))
+			return res.sendStatus(403);
+
+		const updatedPair = await Pants.findByIdAndUpdate(id, pairData, {new: true});
 		res.json(updatedPair);
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
@@ -70,14 +83,20 @@ router.put('/:id', async (req, res) => {
 
 // Delete
 router.delete('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
-		const deletedPair = await Pants.findByIdAndRemove(id);
-		if (!deletedPair) {
+		const pair = await Pants.findById(id);
+
+		if (!pair) {
 			const err = new Error('Pants Not Found.');
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!pair.owner.equals(user._id))
+			return res.sendStatus(403);
+
+		await Pants.findByIdAndRemove(id);
 		res.send('Successfully deleted.');
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
