+++ routes/api/v1/shirts.js
@@ -4,8 +4,9 @@ const	express = require('express'),
 
 // Index
 router.get('/', async (req, res) => {
+	const { user } = req;
 	try {
-		const shirts = await Shirt.find({});
+		const shirts = await Shirt.find({owner: user._id});
 		res.json(shirts);
 	} catch (err) {
 		console.log(err);
@@ -15,6 +16,7 @@ router.get('/', async (req, res) => {
 
 // Detail
 router.get('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
 		const shirt = await Shirt.findById(id);
@@ -23,6 +25,9 @@ router.get('/:id', async (req, res) => {
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!shirt.owner.equals(user._id))
+			return res.sendStatus(403);
+
 		res.json(shirt);
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
@@ -36,9 +41,11 @@ router.get('/:id', async (req, res) => {
 
 // Create
 router.post('/', async (req, res) => {
+	const { user } = req;
 	try {
-		const { shirt } = req.body;
-		const newShirt = await Shirt.create(shirt);
+		const { shirt: shirtData } = req.body;
+		shirtData.owner = user._id;
+		const newShirt = await Shirt.create(shirtData);
 		res.json(newShirt);
 	} catch (err) {
 		console.log(err);
@@ -48,15 +55,21 @@ router.post('/', async (req, res) => {
 
 // Update
 router.put('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
-		const { shirt } = req.body;
-		const updatedShirt = await Shirt.findByIdAndUpdate(id, shirt, {new: true});
-		if (!updatedShirt) {
+		const { shirt: shirtData } = req.body;
+		const shirt = await Shirt.findById(id);
+		
+		if (!shirt) {
 			const err = new Error('Shirt Not Found.');
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!shirt.owner.equals(user._id))
+			return res.sendStatus(403);
+
+		const updatedShirt = await Shirt.findByIdAndUpdate(id, shirtData, {new: true});
 		res.json(updatedShirt);
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
@@ -70,14 +83,20 @@ router.put('/:id', async (req, res) => {
 
 // Delete
 router.delete('/:id', async (req, res) => {
+	const { user } = req;
 	const { id } = req.params;
 	try {
-		const deletedShirt = await Shirt.findByIdAndRemove(id);
-		if (!deletedShirt) {
+		const shirt = await Shirt.findById(id);
+
+		if (!Shirt) {
 			const err = new Error('Shirt Not Found.');
 			err.name = 'NotFound';
 			throw err;
 		}
+		if (!shirt.owner.equals(user._id))
+			return res.sendStatus(403);
+
+		await Shirt.findByIdAndRemove(id);
 		res.send('Successfully deleted.')
 	} catch (err) {
 		if (err.name === 'CastError' || err.name === 'NotFound') {
